+++ res/code3.t
@@ -26,7 +26,6 @@ BOOL1
 SCOLON
 DO
 OB
-OB
 TYPE4
 ID6
 ASSIGN3
