+++ src/com/github/oxaoo/lexer/syntax/Grammar.java
@@ -3,48 +3,6 @@ import com.github.oxaoo.parser.CSyntaxTreeNode;
 
 import java.util.*;
 
-class Symbol {
-    public final String id;
-
-    Symbol(String id) {
-        this.id = id;
-    }
-
-    @Override
-    public String toString() {
-        return id;
-    }
-
-    @Override
-    public int hashCode() {
-        for (String type: Grammar.reservedTypes) {
-            if (id.startsWith(type)) {
-                return type.hashCode();
-            }
-        }
-        return id.hashCode();
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (obj == null)
-            return false;
-        if (this == obj) {
-            return true;
-        }
-        Symbol tr = (Symbol) obj;
-        if (tr != null) {
-            for (String type: Grammar.reservedTypes) {
-                if (id.startsWith(type) && tr.id.startsWith(type)) {
-                    return true;
-                }
-            }
-            return tr.id.equals(this.id);
-        }
-        return false;
-    }
-}
-
 class NotTerminal extends Symbol {
     NotTerminal(String id) {
         super(id);
